requirements.txt
gunicorn==22.0.0
packaging==24.0
sqlparse==0.5.0
tzdata==2024.1
crispy-bootstrap5==2024.2
Django==5.0.4
django-crispy-forms==2.1
gunicorn==22.0.0
packaging==24.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==2024.2
cryptography==42.0.5
cssbeautifier==1.15.1
django-crispy-forms==2.1
djlint==1.34.1
EditorConfig==0.12.4
gunicorn==22.0.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.7
jsbeautifier==1.15.1
json5==0.9.25
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
pycparser==2.22
pytube
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==2024.2
cryptography==42.0.5
cssbeautifier==1.15.1
Django==5.0.4
django-crispy-forms==2.1
djlint==1.34.1
EditorConfig==0.12.4
filelock==3.14.0
fsspec==2024.3.1
gunicorn==22.0.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
jsbeautifier==1.15.1
json5==0.9.25
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
pycparser==2.22
pytube==15.0.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
sympy==1.12
tbb==2021.12.0

typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
crispy-bootstrap5==2024.2
cryptography==42.0.5
cssbeautifier==1.15.1
daphne==4.1.2
defusedxml==0.8.0rc2
Django==5.0.4
django-ckeditor==6.7.1
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-js-asset==2.2.0
django-summernote==0.8.20.0
django-taggit==5.0.1
djlint==1.34.1
easy-thumbnails==2.8.5
EditorConfig==0.12.4
filelock==3.14.0
fsspec==2024.3.1
gunicorn==22.0.0
html-tag-names==0.1.2
html-void-elements==0.1.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
intel-openmp==2021.4.0
Jinja2==3.1.4
jsbeautifier==1.15.1
json5==0.9.25
Markdown==3.6
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
psycopg==3.1.18
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-decouple==3.8
python3-openid==3.2.0
pytube==15.0.0
PyYAML==6.0.1
redis==5.0.4
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
service-identity==24.1.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.4
sqlparse==0.5.0
sympy==1.12
tbb==2021.12.0
Twisted==24.3.0

txaio==23.1.1
tzdata==2024.1
urllib3==2.2.1
webencodings==0.5.1
Werkzeug==3.0.2
google-generativeai
zope.interface==6.4
